Preliminary working (untested) auto, night before comp 😬😬

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/behaviorControl/BehaviorTrees/BehaviorTreeComponents/ActionFunctions/Shooter/SpinUpShooter.java
@@ -24,7 +24,7 @@ public class SpinUpShooter implements ActionFunction {
             return lastStatus;
         }
 
-        targetVelocity = (int) blackBoard.getValue("TargetShooterRPM");
+        targetVelocity = (double) blackBoard.getValue("TargetShooterRPM");
 
         // Activate the shooter mechanism
 //        shooterController.spinToTargetVelocity(targetVelocity);
